requirements.txt
discord.py
toml